fix(management): resolve oidc group mapping
fix gravitee-io/issues#5686

+++ gravitee-rest-api-service/src/main/java/io/gravitee/rest/api/service/impl/MembershipServiceImpl.java
@@ -32,7 +32,6 @@ import io.gravitee.repository.management.api.MembershipRepository;
 import io.gravitee.repository.management.api.search.ApiCriteria;
 import io.gravitee.repository.management.api.search.ApiFieldExclusionFilter;
 import io.gravitee.repository.management.model.Audit;
-import io.gravitee.repository.management.model.Membership;
 import io.gravitee.rest.api.model.*;
 import io.gravitee.rest.api.model.api.ApiEntity;
 import io.gravitee.rest.api.model.common.Pageable;
@@ -130,13 +129,14 @@ public class MembershipServiceImpl extends AbstractService implements Membership
             new MembershipMember(memberId, null, memberType),
             new MembershipRole(roleToAdd.getScope(), roleToAdd.getName()),
             source,
-            true
+            true,
+            false
         );
     }
 
     @Override
     public MemberEntity addRoleToMemberOnReference(MembershipReference reference, MembershipMember member, MembershipRole role) {
-        return _addRoleToMemberOnReference(reference, member, role, DEFAULT_SOURCE, true);
+        return _addRoleToMemberOnReference(reference, member, role, DEFAULT_SOURCE, true, false);
     }
 
     private MemberEntity _addRoleToMemberOnReference(
@@ -144,7 +144,8 @@ public class MembershipServiceImpl extends AbstractService implements Membership
         MembershipMember member,
         MembershipRole role,
         String source,
-        boolean notify
+        boolean notify,
+        boolean update
     ) {
         try {
             LOGGER.debug("Add a new member for {} {}", reference.getType(), reference.getId());
@@ -162,13 +163,19 @@ public class MembershipServiceImpl extends AbstractService implements Membership
                         reference.getId(),
                         optRole.get().getId()
                     );
+
                     if (!similarMemberships.isEmpty()) {
-                        throw new MembershipAlreadyExistsException(
-                            member.getMemberId(),
-                            member.getMemberType(),
-                            reference.getId(),
-                            reference.getType()
-                        );
+                        if (update) {
+                            UserEntity userEntity = findUserFromMembershipMember(member);
+                            return getUserMember(reference.getType(), reference.getId(), userEntity.getId());
+                        } else {
+                            throw new MembershipAlreadyExistsException(
+                                member.getMemberId(),
+                                member.getMemberType(),
+                                reference.getId(),
+                                reference.getType()
+                            );
+                        }
                     }
                 }
                 Date updateDate = new Date();
@@ -504,6 +511,17 @@ public class MembershipServiceImpl extends AbstractService implements Membership
         MembershipMemberType memberType,
         String memberId
     ) {
+        deleteReferenceMemberBySource(referenceType, referenceId, memberType, memberId, null);
+    }
+
+    @Override
+    public void deleteReferenceMemberBySource(
+        MembershipReferenceType referenceType,
+        String referenceId,
+        MembershipMemberType memberType,
+        String memberId,
+        String sourceId
+    ) {
         try {
             Set<io.gravitee.repository.management.model.Membership> memberships = membershipRepository.findByMemberIdAndMemberTypeAndReferenceTypeAndReferenceId(
                 memberId,
@@ -513,9 +531,11 @@ public class MembershipServiceImpl extends AbstractService implements Membership
             );
             if (!memberships.isEmpty()) {
                 for (io.gravitee.repository.management.model.Membership membership : memberships) {
-                    LOGGER.debug("Delete membership {}", membership.getId());
-                    membershipRepository.delete(membership.getId());
-                    createAuditLog(MEMBERSHIP_DELETED, new Date(), membership, null);
+                    if (sourceId == null || membership.getSource().equals(sourceId)) {
+                        LOGGER.debug("Delete membership {}", membership.getId());
+                        membershipRepository.delete(membership.getId());
+                        createAuditLog(MEMBERSHIP_DELETED, new Date(), membership, null);
+                    }
                 }
             }
         } catch (TechnicalException ex) {
@@ -1366,7 +1386,7 @@ public class MembershipServiceImpl extends AbstractService implements Membership
             }
             return roles
                 .stream()
-                .map(role -> _addRoleToMemberOnReference(reference, member, role, source, notify))
+                .map(role -> _addRoleToMemberOnReference(reference, member, role, source, notify, false))
                 .collect(Collectors.toList());
         } catch (TechnicalException ex) {
             LOGGER.error("An error occurs while trying to update member for {} {}", reference.getType(), reference.getId(), ex);
@@ -1384,6 +1404,9 @@ public class MembershipServiceImpl extends AbstractService implements Membership
         Collection<MembershipRole> roles,
         String source
     ) {
-        return roles.stream().map(role -> _addRoleToMemberOnReference(reference, member, role, source, false)).collect(Collectors.toList());
+        return roles
+            .stream()
+            .map(role -> _addRoleToMemberOnReference(reference, member, role, source, false, true))
+            .collect(Collectors.toList());
     }
 }
